apps/api/requirements.txt
quart==0.20.0
python-dotenv==1.0.1
pydantic==2.9.1
openai>=1.0.0
httpx>=0.27.0
PyJWT>=2.8.0
azure-cosmos>=4.7.0
certifi>=2024.0.0
